backend/requirements.txt
torch
torchvision
pillow
flask
flask-cors
